moved SearchLoader to its own file
