+++ src/client/components/searchLoader/index.js
@@ -1,0 +1,3 @@
+import SearchLoader from './SearchLoader';
+
+export default SearchLoader;
